+++ src/caselawclient/models/utilities/dates.py
@@ -1,15 +1,14 @@
-from datetime import datetime
+from datetime import UTC, datetime, tzinfo
 
 from dateutil.parser import isoparse
-from pytz import UTC, tzinfo
 
 
-def parse_string_date_as_utc(iso_string: str, timezone: tzinfo.BaseTzInfo) -> datetime:
+def parse_string_date_as_utc(iso_string: str, timezone: tzinfo) -> datetime:
     """iso_string might be aware or unaware:
     ensure that it is converted to a UTC-aware datetime"""
 
     mixed_date = isoparse(iso_string)
-    aware_date = mixed_date if mixed_date.tzinfo else timezone.localize(mixed_date)
+    aware_date = mixed_date if mixed_date.tzinfo else mixed_date.replace(tzinfo=timezone)
 
     # make UTC
     utc_date = aware_date.astimezone(UTC)